Fix critical player data filtering bug + add diagnostic
PROBLEM:
Player data was being filtered to 0 rows during window training,
causing "Loaded 0 player-games for window" error. This made player
prop models completely non-functional.

ROOT CAUSE:
Type mismatch in season filtering (train_auto.py lines 5018, 5040):
- _season_from_date() returns float64 (numpy.where output)
- padded_seasons is a set of Python ints
- .isin() comparison failed silently, filtering out ALL rows

FIX (train_auto.py):
1. Line 5018: Add .astype('Int64') to season calculation
2. Line 5040: Add .astype('Int64') to season calculation

This ensures type consistency for .isin() filtering.

DIAGNOSTIC TOOL (diagnose_player_filter.py):
New script to test player data loading and filtering BEFORE
full training. Helps identify:
- Date parsing issues
- Season calculation problems
- Type mismatches
- Memory problems

Run in Colab: python3 diagnose_player_filter.py

FUZZY MATCHING:
Already exists in train_auto.py (lines 2587-2703).
Name normalization handles:
- Unicode/accents
- Suffixes (Jr., Sr., III)
- Punctuation
- Capitalization
Season offset matching (±1 year) for player priors.

+++ train_auto.py
@@ -5015,7 +5015,7 @@ def main():
                 # Filter historical to window immediately (before concat to save memory)
                 date_col = [c for c in hist_players_df.columns if 'date' in c.lower()][0]
                 hist_players_df[date_col] = pd.to_datetime(hist_players_df[date_col], errors="coerce")
-                hist_players_df['_temp_season'] = _season_from_date(hist_players_df[date_col])
+                hist_players_df['_temp_season'] = _season_from_date(hist_players_df[date_col]).astype('Int64')
                 padded_seasons = set(window_seasons) | {start_year-1, end_year+1}
                 hist_players_df = hist_players_df[hist_players_df['_temp_season'].isin(padded_seasons)].copy()
                 hist_players_df = hist_players_df.drop(columns=['_temp_season'])
@@ -5037,7 +5037,7 @@ def main():
                 # Filter to window
                 date_col = [c for c in hist_players_df.columns if 'date' in c.lower()][0]
                 hist_players_df[date_col] = pd.to_datetime(hist_players_df[date_col], errors="coerce")
-                hist_players_df['_temp_season'] = _season_from_date(hist_players_df[date_col])
+                hist_players_df['_temp_season'] = _season_from_date(hist_players_df[date_col]).astype('Int64')
                 padded_seasons = set(window_seasons) | {start_year-1, end_year+1}
                 hist_players_df = hist_players_df[hist_players_df['_temp_season'].isin(padded_seasons)].copy()
                 hist_players_df = hist_players_df.drop(columns=['_temp_season'])
